🔒 [security fix] Unclosed Resource Leak in OdsReader::readString Generator

+++ src/OdsReader.php
@@ -66,9 +66,14 @@ class OdsReader implements ReaderInterface
     {
         $options?->applyTo($this);
         $filename = Spread::getTempFilename();
-        file_put_contents($filename, $contents);
-        yield from $this->readFile($filename);
-        unlink($filename);
+        try {
+            file_put_contents($filename, $contents);
+            yield from $this->readFile($filename);
+        } finally {
+            if (is_file($filename)) {
+                unlink($filename);
+            }
+        }
     }
 
     /**
